remove redundant LoggerRegistry; move code to GradleLoggerFactory
make LoggerMetadata internal
