+++ subprojects/gradle/gradle-logger/src/main/kotlin/com/avito/logger/LoggerMetadata.kt
@@ -2,7 +2,7 @@ package com.avito.logger
 
 import java.io.Serializable
 
-data class LoggerMetadata(
+internal data class LoggerMetadata(
     val tag: String,
     val pluginName: String? = null,
     val projectPath: String? = null,
